Created MapCursor class.
This is a pretty large commit that adds the completed (for now)
mapCursor functionality to the game. Looking back on it, I should've
committed as I went, but it's too late for that now.

A detailed explanation of the MapCursor's functionality and logical
implementation is included at the top of the file, and detailed
documentation on specific features in included throughout the rest
of the file.