Added temporary code to MenuState for testing.
The one important-to-note feature is the addition of mapCursor.update
to the public update function. Since mapCursor isn't a child of an updating
HaxeFlixel class, it needs to be manually updated by the scene itself.

Its order of updating (either before or after everything else) will cause some
behaviors to change slightly (mostly determining which game object "wins"
ties if the player tries to move the cursor and press an action button on the
same frame) but placing the update in either position shouldn't break anything.