Created MoveInputHandler class.
This class is a pretty close copy of the movement logic from MapCursor,
but is generalized a bit and implemented with static methods so any class
can access it as needed. My reasoning for this decision is explained in the
comments at the top of the class, but it basically boils down to this: It
seemed like the best way to make this movement logic available to
any future game components that will need it.

In the following commit I will remove MapCursor's old movement logic and
set it up to interact with the new system.