Created BoxCreator class.
BoxCreator is a class with all static fields (essentially a singleton) that
can be used to create (and return!) FlxSprites of boxes of any size,
constructed using the graphics selected in its setBoxType function.

The details of box-creation logic is laid out in the comments at the top of
the file, as well as a specification for the format of the box spritesheets
that it uses to create the boxes.