Created HideableEntity interface.
Like the updating interface, this isn't really intended for or needed by any
HaxeFlixel-inheriting classes, since can pretty easily be hidden by flipping
a boolean variable. This is instead meant for classes that are responsible
for managing multiple visual components, that need a function to hide
(and another to reveal) each of those visual components.