Renamed CornerSprite to AnchoredSprite.
This was necessary because I realized that the movement logic of
CornerSprite was applicable and useful for a much wider variety of
sprites than just corners, such as a bouncing cursor for a menu. Since
the "anchored" component of the CornerSprite was the core of how its
movement worked, I decided to rename it to AnchoredSprite, a class
that could be used to make bouncing corners, cursors, or anything
else that needs anchoring.

Also moved the sprite-flipping decision-making out of AnchoredSprite
and into MapCursor, since most AnchoredSprites won't be concerned
with flipping vertically or horizontally.