Added InputEvent typedef and EventExtender static extension.
These two files work together to build the foundation of an event system
 that will eventually be used by the Observer design pattern. I'll go into
the Observer pattern and why it's being used in a later commit, don't worry.
For now, we just need to remember that an event system will be nice to have
when we get to implementing the Observer pattern.

Anyway, the InputEvent typedef basically just sets up a new alias for an Int,
so instead of saying Int, I can say InputEvent. On its own, this is pretty boring,
but what EventExtender does is create a set of functions that the InputEvent
type can use to do stuff that an ordinary Int can't do. In this case, it allows
InputEvents to pack two different 16-bit integers inside itself instead of just
being a single 32-bit integer.

Yes, I could just make an input event class that has a couple of public integer
variables and achieve the same sort of thing, but this is neat because all of the
data is packed into a single 32-bit integer.

I used a typedef and static extension pair similar to this in my Potion-Shop game
for its event system and it worked great, so I figured I might as well do it again here.