Implemented the rest of the Observer Pattern.
The Subject class created for this project will be used to allow
classes with a limited view and/or knowledge of the game state around
them to cause changes like hiding or revealing other menus, making
characters move around the game, and more.

The whole idea is that when an Observed-type class detects some sort of
input event that needs to cause a change in some external part of the game,
it just uses a Subject object it controls to notify that input event to whatever
objects may be listening.

When an Observer-type object that has access to a larger part of the game
state observes a notification, it uses the data from the notification (includes
the id of the subject that did the notification, the type of event that occurred,
and the parent object of the notifying subject) to cause some larger change
in the game as a whole.

The InputEvent typedef and EventExtender static extension are useful in this
pattern for providing information to the Observer that isn't otherwise tracked
by the observed class. After an input event occurs, the Observed-type classes
won't necessarily have some variable that keeps track of what sort of input
event just occurred.

This sort of Observer Pattern strategy was used in Potion-Game to good effect,
and I feel like it'll be an effective tool here as well.