Adjusted MoveInputHandler timeHeld logic.
Under the new system, timeMoveHeld is only increased on the
first call to attemptHeldMovement, and is not changed on any subsequent
calls. It still isn't a great idea to have multiple objects trying to use one
inputHandler at the same time, but at least the behavior for all of the
movement should generally work correctly even if that is the case.