Created ActionInputHandler class.
Overall, this class is very similar to MoveInputHandler (to the
point that I may want to go back and have them inherit from
some common parent) but is used to check for "action"
inputs rather than movement.

At the moment, those inputs include:
Confirm - used to select options and indicate an affirmative choice.
Cancel - used to exit menus and indicate a negative choice.
Paint - (eventually) used to mark units on the map to help see their
     attack range.
Next - used to cycle through characters on the battlefield.
Info - used to pop up a help window for complex menus/info screens.

Unlike the movement handler, though, this class has the option of
not attempting held action inputs, and will only check for them if
the caller to handleActions() specifically indicates to attempt them.