Added ActionInput responses and new interfaces to MapCursor.
This commit adds the newly created interfaces to the MapCursor class,
along with all of the variables and functions required for them to work.
As a result, MapCursor is now a HideableEntity, an UpdatingEntity, and
an Observed object. As a result of being Observed, it has a Subject now,
and can notify its observers of new events.

It also adds integration with the ActionInputHandler class. It responds
to action button presses by notifying any observers of the event, and
(in the case of the confirm button) playing a sound effect.

I guess these changes aren't directly tied to the implementation of a basic
menu, but they were changes that became possible because of the other
advances I made, which were needed to get a solid basic menu set up.