Created MenuOption class.
This class is the basic unit of all future menus. Each menu option has
a point for the menu cursor to anchor to, a label for the menu option,
a background highlight to indicate when the MenuOption is hovered over,
and a set of references for each direction, to indicate what MenuOptions
neighbor this one for movement purposes.

A simple menu will create a set of these MenuOptions, some sort of
background box graphic using the BoxCreator class, and a menu
cursor to move between these.

MenuOption may end up being extended to do more in the future
(the background highlight is very bare-bones right now) but it is
definitely more than serviceable enough for this project.