Created MenuTemplate class.
This class contains the functionality that I expect all menus to need.
As stated at the top of the file, this includes:
- Containing MenuOptions.
- Hiding menu and all components.
- Activating/Deactivating menu.
- Reacting to user "action" input.
- Uses observer pattern-style system, using its Subject component to notify observers.
   - Cursor creation.
   - Includes side-to-side bouncing motion and moving between menu options.
- Cursor movement.
   - Includes tracking the current menu option for movement and action input purposes.
- Menu movement.
   - Includes moving all menu components to match.

If it turns out that certain functionality is only used by a certain set of menu objects,
then I can always move those functions into a child class that acts as a more specific
template for that set of menus.

It's important to note that this class isn't usable as a standalone class. It must be
extended by a child class to actually put its functions to use in an actual menu.