Created BasicMenu class.
Unlike the MenuTemplate class, this one is a functional and usable
menu! It's meant to be used for the set of popup menus that are used
while on the battlefield. That type of menu just has a vertical list of options
that the player can select from which should have a variety of effects,
which often includes popping up other menus.

The idea behind this menu is that a list of strings (the options in the menu)
are specified during its instantiation, and it will automatically construct
MenuOptions with the correct cursor anchor coordinates, label text, &
background highlight width as well as a correctly sized background
box using the BoxCreator class. Instead of building each of these simple
menus independently with hand-selected widths and cursor locations,
the menus are automatically created according to a standard set of rules.

These menus are completely set up to integrate with the
MovementInputHandler and ActionInputHandler classes, and implement
the Observed interface so they can notify external classes when menu
options are confirmed or canceled.

Overall, I'm very happy with how this class turned out!