Added new functionality to MenuState for testing purposes.
Everything in here is very rough, and not meant to be any sort of
representation of how the final game will be put together. It's
just a place for me to throw together code to do some quick
testing, and as a way for me to demo the most recent changes
to the game.

In this commit, menus and the Observer pattern are on display.

To open the the first menu, press "Z" (the confirm key).
To open the next menu, press "Z" again, etc.
To close the current menu and return to controlling the MapCursor,
    press "X" (the cancel key).

While controlling the map cursor, any of the action keys (z, x, c, s, d)
can be pressed to cause some text to be written to the console describing
what sort of event occurred.

When menus are open, pressing "Z" or "X" will cause a similar message to
be written to the console. Additionally, pressing "Z" will write a message to
the console stating which menu option the player selected by pressing "Z".

Again, everything here is very rough, and is not meant to be permanent in
any way.