Moved Observer Pattern related files to a new set of folders.
Also updated package and references where necessary.

Was done because "utilities" is a pretty vague folder name,
and I'll do my best to move source code files to more specific
folders when possible to improve organization and make it
easier to reuse specific components of this project in future ones.