Created cameraHitbox component.
To make it possible for the game camera to correctly track the movement of the
map cursor, it needed to keep track of a tile-sized FlxObject that moved along with
the cursor's corners, but always remained in the cursor's "actual" position, rather
than the bouncing around or strangely offset positions of the corner sprites.

To accomplish this, the cameraHitbox FlxObject was created.
The cursor's jumpCorners function was changed to also move the cameraHitbox
the same amount as the corners, and was renamed to jumpCursor to better reflect
how it moved the entire cursor, not just the corner sprites.

I also created the updateCameraHitboxPos() function, which takes the anchor position
of the top-left corner and subtracts the anchor offsets to get the x & y values that the
cameraHitbox should have, allowing it to smoothly follow the corners as they move
from tile to tile.

IMPORTANT NOTE:
updateCameraHitboxPos() must be called ***before*** the call to moveCornerAnchors,
unless we change the order of update calls & super.update() in MissionState.
See the comment above update() for a thorough explanation of why this is the case.