Added Ogmo Editor files to game project.
The forest_1.oel is a level file used by the game to lay out a tilemap and everything
that involves.

I created the "other" folder for the .oep file because it didn't seem
correct to put it in the game's asset files. It's the "Ogmo Editor Project"
file, used to load the game's information into the Ogmo Editor, and nothing
else.

I think it's nice to have with the game project as a whole, so myself or anyone
else cloning down the project will be able to open up Ogmo Editor and change
a level or create a new one using the same project settings I used on my computer.

Side note: when you open the project, it changes the filename line of the .oep file.
There's no need to commit that change to the repo in the future.