Created MissionState & set as default game state.
At this point, MissionState is basically the same as MenuState. Making a
new state was necessary, however, because this branch involves setting
up a camera and a map, both of which are going to be final elements
in the mission sections of the game. Rather than do a bunch of work
in a test environment and try to move it to a new state later, I just set
up the MissionState now as the target for development related to
the map and camera.