Changed BasicMenu sprites to have scrollFactor of (0, 0).
This was done so that menus pop up on the game screen in the same
positions regardless of where the camera has scrolled off to.

To implement this feature, a setScrollFactors() function was added to
MenuTemplate, which BasicMenu used in its initialization function.
setScrollFactors() could not be called inside MenuTemplate's new()
function because it doesn't actually instantiate any of a menu's final
visual components.

As a result, it's up to the MenuTemplate's child classes to use
setScrollFactors if they want to have the menu position be the same
regardless of the camera position.