Added input buffering system to ActionInputHandler.
I realized that MapCursor doesn't necessarily want to respond
to action inputs on every frame. When moving between tiles,
any new inputs should be placed in a buffer, which can be
processed when the cursor arrives.

This is necessary because just responding to buttons when they're
pressed means that a menu could pop up while the cursor was
halfway through a move, which looks particularly bad if its moving
the camera along with it (we'd get lovely half-tiles at the edge of the
screen).

This does change the way that external objects interact with the
ActionInput handler in a few ways. ActionInputHandler's bufferInputs
function must now be called at the start of its state's update function,
after which external classes are free to call useBufferedInput, which
consumes a single input event. If no classes call useBufferedInput
for several frames, the buffered inputs will stack up, waiting to be
used.

Honestly, it's a system that's semi-complicated that will have very
little visible effect on the game (the window where the cursor can't
accept actions is only 4 frames long right now) but I think it is an
important enough component of getting the game's feel right that
it's worth putting together.