Updated MapCursor to use the new input buffer system.
In the process of doing this, I added a boolean variable that keeps track
of whether the cursor has arrived at its destination or not, which made a
couple of spots of code more readable.

The update function is a bit more complicated now, since any one of the
multiple (potential) buffered actions could change the MapCursor's
currInputMode to DISABLED, which means we wouldn't want to
have it read any more action inputs from the buffer or process
any movement inputs.