Added strategic tile layer to the Ogmo project and level.
The idea here is that there may be many different visual tiles for a particular
type of terrain, but in terms of strategic/tatictical benefit, they're identical. To
express that in the Ogmo project and level, I added another tile layer to the
project that won't ever be visually present in the game, but will be used to
generate a "terrain map" that can be used to determine what sorts of
tactical bonuses characters should have based on their column/row
position on the map.

As of right now, though, there is no pre-built way for me to get the data I want
out of the ogmo level file, so I'm going to have to do a bit of extending of the
FlxOgmoLoader class.