Created StrategyOgmoLoader class.
As mentioned in a previous commit, adding a "strategic" tilemap
to the game's levels would require some extra level-loading functionality
to take advantage of. This extension of the FlxOgmoLoader class adds
a pair of functions for the purpose of getting, parsing, and storing the data
held in the level's strategic tile layer.

There are some assumptions at work here, that certain properties
will exist in the level's xml file. If there is no "number_of_terrain_types"
or no set of "terrain_type_tile_*" properties in the level, the game will
crash when the level is loaded.