Created MapCursorManager and integrated it into MissionState.
MapCursorManager is the first of several managers that will be added to the game.
The idea behind them is to encapsulate different parts of MissionState's responsibilites
in small classes composed within MissionState. These small managers will have access
to most of MissionState's data fields, and will be able to call MissionState's public facing functions.

Under this strategy, MissionState will largely serve as an interface for the various managers to interact
with, which will hopefully keep the code much cleaner than it would be if all of the managing functionality
was just kept inside MissionState.