Added functions to create bottom/right sides and corner graphics.
These functions are needed to implement dynamically resizeable menus,
which will be needed for the battle menus that are the primary focus of this branch.

They'll be put to use in the next commit, so check that out if you're not sure how they
can be used to make a dynamically resizeable box.