Changed visibility properties of VarSizedBox variables.
If you're looking for the implementation of a resizeable box,
look at the next commit. Sorry!

This visibility of all VarSizedBox properties were changed so that external
classes can view box properties, but not change them directly. There never
was a situation where an external class would try and change box properties,
but I figured I might as well enforce it with the visibility stuff.