Created ResizeableBox class.
This class creates and maintains a resizeable box.
It should be used in a situation where a box's size should change depending different
context-sensitive variables. Generating a bunch of boxes for every concievable situation
could work, but that would be a pretty wasteful solution by comparison.

An example of a use case for a ResizeableBox is a battle menu that only displays options
when they can be executed by the selected unit, e.g. the "attack" option only appears if the
selected unit is in range of an enemy. Instead of making a separate menu for every possible
permutation of the available options or having unselectable options greyed out, this resizeable
box allows the menu size to change at will.

The box is resizeable due to the new functions added to BoxCreator.

For the box to become smaller, it clips the "maximum-sized" box image generated
during initialization down to the newly-selected dimensions. Clipping the graphic doesn't
completely solve the resizing problem, though, because all it does is stop displaying the
bottom and right sections of the box that are outside of the clipping rectangle.

Since the box is just created from a bunch of small component graphics, we can solve this
by using a separate sprites that are just the bottom & right border sections of the box. The
edge sprites are displayed over the bottom and right edge of the clipped "maximum-sized"
box sprite, and the box regains its bottom and right borders.

At that point, there's only one more problem to solve. If the bottom and right border sprites
are always just the same length/height of the orginal max-sized box, then their borders stick
out past the box on the bottom and right sides. To solve this, we also clip the side rectangles,
so they don't extend past the box, then add a sprite of the bottom-right corner of the box over
the bottom-right corner area to hide the cut off edges of the two borders.