Changed when menu components were added to totalFlxGrp.
The order of adding FlxSprite-inheriting classes to a menu's totalFlxGrp matters,
since it determines which sprites appear on top of/beneath others when the
group of sprites is displayed.

This change creates a dedicated function for adding components to the totalFlxGrp,
so the order of adding to totalFlxGrp can be properly specified there, instead of being
scattered across a number of different functions.