Made MenuOption implement HideableEntity.
It's convenient to be able to hide or reveal individual menu options
at will in some cases, and adding these functions makes it possible
to do just that.

I'm starting to think that I should just put together a few static functions on
a hideable entity class, though, so I don't have to have the exact same set
of functions written out in every class that needs to be able to hide its components.
I may make a change to that effect at some point later in this branch.