Fixed spelling error in file name, and misc. fixes & changes.
Sorry for having a general commit, but I forgot that when renaming
files you have to commit all changes to those renamed files in one go.
I'll make sure I only rename files when no other changes are going on.

Made ResizableBox's x & y variables publicly accessible and alterable.

This is so that the ResizableBasicMenu can manually change the box's x & y values
during its setPos() function, since the box's logical position needs to match the
new position of its graphical components.

Made ResizableBasicMenu's MenuOptions change IDs when changeMenuOptions()
is called.

This change is needed because previously the MenuOption ids weren't very useful. All
they told you was number option they were in the overall menu, with no hint at what
that menu option was representing at that moment.

With the new system, the MenuOption's IDs match the id number that the label text was
originally associated with. So if option # 5, "run" is changed to be displayed in the # 0 spot,
the # 0 menuOption's ID will be 5, since its acting like option # 5 (for now). There's a different
explanation at the top of the ResizableBasicMenu file if you'd prefer to see that one instead.