Added region-tracking functionality to MapCursorManager.
It now has a pair of boolean variables that signify what half of the screen the
map cursor is currently on, one for left/right, the other top/bottom. External
classes can use these variables for plenty of things, the main one being
positioning menus on the opposite side of the screen from wherever the
MapCursor is.