Created PossiblePosTracker class.
The idea behind this class is to bundle together data about the positions
an object could be in. Tracks 4 variables, one for each of left & right x values
and bottom & top y values.

Will be used to remember positions for menus that need to switch between
left and right sides of the screen, and other objects that need to function in a
similar manner.