Instated correct MissionState functionality.
This structure is far more in line with what the MissionState will
actually be doing in the final version of the game. It has public-
facing functions that can be called by its managers to facilitate
communication between them.

Documentation should, for the most part, explain what's going
on pretty well. I may go in and add in some more detailed
explanations of the design later if it feels like theyr'e necessary.