Adjusted MapCursor movement logic.
Previously, the MapCursor moved a set number of pixels every frame, which was determined by
taking the tilesize / number of frames per move. This system was a bit inconvenient, because the
number of frames per move had to evenly divide the tilesize.

Under the new system, we can use any number as the frames per move, and I changed it to be
6 frames rather than 4. To make up for the slower speed, I slightly reduced the delay on held
movement.