Created MoveID type and MoveIDExtender class.
This pair functions similarly to the InputEvent and EventExtender. It's main function is to bundle
multiple pieces of data into a single primitave variable. In this case, row/column data that describes
a particualr tile in the game's map can be stored inside MoveID.

The MoveIDExtender provides functions to get the Row and Column info from a MoveID, as well as do
distance and direction calculations between two different tiles.