Created TileChange class.
TileChange objects are used to track changes that have happened to the map as a result of
units moving. A unit cannot move through a hostile unit, so when a hostile unit takes a new position
its old spot needs to be noted as newly opened, and its new spot needs to be noted as newly closed.
This class has all of the variables needed to describe these changes fully.

Will be used in the upcoming UnitManager class.