Created ArrowTile class.
Fairly straightforward extension of the FlxSprite class, used to draw the
path of movement a unit will take to move to a particular square. Uses
single-frame animations to switch between the different arrow-line segments
that may need to be used to draw a line. The animations are named in
a pattern that makes it easy to use them programmatically, see the
documentation inside the file for details.