Created initial version of many Unit-related classes.
Most of these classes are basically skeletons at the moment.
I plan to fill them out later when I start delving deeper into unit mechanics.
For now, I'll commit all of these so I have a basic, placeholder version of
a unit to play around with.

Also, documentation in these files is pretty lacking, but that'll be fixed before
a full PR happens.