Created the MapCursorUnitTypes enum.
Very straightforward enum. Put together so I had a convenient way of
organizing the units according to what matters to the MapCursor.
The Mapcursor cares far more about the active/inactive state of a
player unit than the details of whose team the "NOT_PLAYER"
unit is on, for example.