Created UnitManager class.
Like the other Manager-type classes, the UnitManager is responsible for all
of the unit-related actions that needed to be handled in the MissionState.

These unit-related actions include:
 - Creating and placing Unit objects into the game.
 - Tracking the positions of all Unit objects.
 - Calculating the movement & attack ranges of all Unit objects.
    - Performing initial calculations
    - Tracking movements across the map so movement/attack ranges can be
      partially recalculated as necessary.
 - Displaying the movement & attack range of a selected unit.
 - Calculating the movement path of a selected unit based on the movements
    of the MapCursor.
 - Displaying & updating the movement path of the selected unit.

The internal documentation will describe the implementation of each of
these components in detail. There are some undocumented functions
right now, but those will be handled in a future commit.