Adjusted MoveMode & InputMode property usage.
Both were changed to be publicly visible (but only alterable through
in-MapCursor methods). This makes the modes useful as indicators
of the MapCursor's current state, which may be used in conditional
statements in the MapCursorManager.

The entries in the InputModes were also changed, since I realized
there needed to be a distinction between player units being selected
and other units being selected. Specifically, held movements should
only be constrained when in the PLAYER_UNIT input mode (changes
that implement that feature will be added in a future commit).