Restricted held movement when in PLAYER_UNIT inputMode.
One of the components of the game's design is that the mapCursor
should not be allowed to exit a selected player unit's valid movement
tiles via a held movement input. It can still be exited normally via a
pressed movement, but this makes it easier for a player to get to the
particular square of movement they want the unit to reach.

To implement this, the MapCursor now has a "selectedLocations"
public property that will be set by the MapCursor manager to reference
the MoveTiles map of the currently selected unit. This allows the
MapCursor to make its own judgements on whether it should
allow held movement or not when in the PLAYER_UNIT inputMode.