Made cursor immediately change to correct moveMode appearance.
Previously there was a 1-frame delay before a change took place, because
the tweens would be set up on the frame of the change, but wouldn't actually
be used to change the corners' visual positions until the tween functions were
called at the start of the next frame.

Now the function that the tween calls each frame is also called once with initial
setup values whenever the movemode is changed, that way the corners immediately
shift visible position without delay.