MapCursor now notfies a MOVE event when jumping to location.
Previously, this didn't have any impact on the game, so it didn't matter
that a move event was notified. In an upcoming commit, however, the
MapCursorManager keeps track of a prevCursorPos and currCursorPos
which need to be changed every time the MapCursor's position changed,
regardless of whether that change was a traditional 1-tile move or a leap
between distant tiles.

Without this, there ends up being a bug in the unitManager's arrow drawing
behavior that occurs because the prevCursorPos & currCursorPos weren't
being updated after cancelling and reselecting a player-controlled unit.