MapCursor now stores info about the cursor's state before disabling it.
This will eventually be useful for when the cursor is hidden after moving
a player controlled unit. If the player cancels back out of the menu,
the cursor should return to the PLAYER_UNIT state rather than going
back to the FREE_MOVEMENT state. To make this system a bit more
elegant than "handle it on a case-by-case basis," I created a variable
to remember the previous state so the cursor can be (optionally)
activated with the same inputMode that it was orginially deactivated with.