Set up all newly created classes inside the MissionState.
This includes importing the files, creating a variable to hold the UnitManager,
initializing the MoveIDExtender, instantiating the UnitManager, and adding all
UnitManager components to the MissionState.

These components will be used heavily by the code in the upcoming commit(s)
because it's about to start integrating the new unit system with already existing
game systems.