Created variable for the state of the MissionState.
It bothers me that I need to use state to refer to two different things,
but we'll survive, I'm sure. The purpose of this variable was already
discussed in the commit/documentation for the PlayerControlStates
file, but the basic idea is that this variable will track what interaction
state the MissionState is currently in, so it knows how to interpret
various types of input events and function calls that come from its
managers.