Created isMapCursorOverUnit() function.
The name is pretty self explanatory. It uses the position data from
the mapCursorManager to get unit data from the unitManager.
Then it determines what entry of MapCursorUnitTypes that this
unit matches, and returns that type to the caller.

This function should only be called by the MapCursor as of now.