Created calculateNewMoveArrow() function.
Uses previous and current cursor position information from the
mapCursorManager to identify what tile the cursor moved into.
If the cursor moved diagnoally, the move needs to be processed
as a pair of orthagonal moves, since the arrow needs to be drawn
tile-by-adjacent-tile.

In the end, this function just sends off a call to the unitManager's
updateMoveArrow() function.