Added functions for confirm and cancel events from the mapCursor.
This functionality used to be handled inside the MapCursorManager class,
but as the game started becoming more complex, the decision making process
that follows a confirm or cancel event involved more information than what the
MissionState is aware of. As a result, the functions were moved into the MissionState
as a whole.