Added small optimization to breadth-first move range calculation.
A tile is only added to the queue to be processed if  its cost is less than
the unit's move property. If the cost == the unit's move, then there definitely
won't be any new movement tiles reached from that tile, so it doesn't need to
be added to the processing queue.