Added findNeighborPathToTarget function.
This function can be used to find a series of directions a unit should
use when moving to a particular destination within their moveTiles.
A player-controlled unit won't need to use this, since it can just use
the movePath array of ArrowTiles, but units that don't have an arrow
path drawn before their move (all non-player-controlled units) will
need to use this to determine how they should move to their selected square.