Adjusted the cursorManager's strategy for storing cursorState.
The point of storing the old cursor state was that it would be easier to return
the cursor to whatever state it was in before it was disabled. With that in mind,
we don't want the stored cursor state to be Disabled. Instead, we should just
keep the previous "old cursor state".