Integrated menu-closing logic into MissionState's controlState system.
It isn't perfect yet, but it's getting there. The goal is to have the MissionState
get told if the unit did some sort of turn-ending action while in the menu system.
If so, the unit should have its logical position be updated and be unselected.
If not, then the cursor and character should jump back to its old logical position
and pick a new location to move to.