Added a public function to be called when the unit's move ends.
This will be used by the UnitManager when the unit's sprite finishes moving to
its destination tile. At that point the MissionState will be able to actually move
on to opening the unit_action menu.

Note: this should only be done if the unit is player-controlled. Enemy unit's should
just move on to executing the next part of their action.