Added update function to UnitManager.
If the movement-related variables indicate that a unit is in the
process of moving, then this update function makes sure the
proper functions are called to keep the unit scooting along its
path of movement.

On the frame that the movement finishes, this function will
call the parentState's unitMovementFinished() function.