Fixed a pair of bugs affecting blocked move/attack recalculations.
There was some buggy behavior in the game previously when two allied
units moved into adjacent spaces and then you tried to view an enemy
unit's movement (whose move range should be limited by the two units).
If both player units were moved and then  the enemy move range was
checked, the game crashed. If the enemy range was checked between
each player unit move, the game wouldn't crash, but the enemy would
display extra attack tiles that shouldn't be in its attack range any longer.

The extra attack range bug was a really simple fix: I had forgotten that the
concat Array method returns a new array instead of modifying the calling array,
so I just needed to actually assign the results of the function back to the calling array.

To understand the second bug, we must note that a unit's move range recalculation function
takes an array of the moveIDs that are both orthagonally adjacent to one of the blocked
tiles and are still in the unit's movement range (i.e. were not outside of the unit's move
range before the tile was blocked) as a parameter.

The crash happened because when two adjacent tiles were blocked, the second blocked tile
was added to the list of "valid move tiles that are adjacent to a blocked tile" because it was
adjacent to the first tile and hadn't yet been removed from the unit's movement range. Later in
that function, it would be removed from the movement range, and as a result the move recalculation
function's array of moves to check contained one moveID that is no longer in the unit's moveTiles,
resulting in a "null" valued tile that causes a crash in the function.

This was fixed by using a different test function for tileBlockedRecalc()'s call to
getValidNeighbors(), which is called block recalcTest().