Moved cursor- & MenuOption-related code into CursorMenuTemplate.
The old MenuTemplate class wasn't as general as it needed to be.
Rather than acting as a basis for all menus, it contained code related to
a bouncing cursor and menu options, which are only really needed for
a traditional, cursor-controlled style menu.

Now I'm starting to make target-style menus, which don't have a list
of options but  instead cycle through different units when receiving move
inputs.

To avoid rewriting code, in a new MenuTemplate class, I took all code that
wasn't common to both target-style and cursor-style menus and moved it
from MenuTemplate into CursorMenuTemplate. CursorMenuTemplate will
act as an intermediate-level menuTemplate that adds layers of functionality
specific to a particular type of menu without actually being an instantiable
class.

In upcoming commits, I'll begin work on a TargetMenuTemplate class that
serves a similar purpose, defining functionality common to all target-type
menus without specifying anything needed by a particular menu.