Changed null MoveIDs to -1 to fix building to native code.
Static targets don't allow primitives to have a value of Null unless its type is
Null<T>. And even then, it really just sets the value of an int-type primitive
(like my MoveID) to 0.

To make the game buildable on static targets, I changed all spots where
I set a MoveID to null (or compared one to null) with -1, since that is an
"invalid" MoveID value.