Added OnMapEntity interface for all map-existing objects to implement.
The OnMapEntity interface is very simple; it just requires that anything
that implements it has a single MoveID that describes its position on the map.

I changed Unit to implement OnMapEntity, which just involved merging its
mapRow & mapCol properties into the single mapPos property. As a
result of this change, other files needed to be changed to properly use the
Unit's mapPos property.