Added a getValidUnitsInRange function to MissionState and UnitManager.
This provides a public interface for targeting menus (or other classes) to
provide some a set of ranges and a testing function and in return receive all
valid units within that range of the currently selected unit.

Overall, the code is quite similar to getValidNeighbors, but rather than getting
moveIDs, it gets units.